Stop launching dataGathering phase after auth. Add usernameAlias
This PR fixes one bug and adds one new component which is part of
version 2.1.0 of the WDC.

The first change is to stop launching the dataGathering phase after
testing the auth phase. This is necessary to properly simulate adding
user credentials independently of a datasource as we do on Tableau
Server.

The other change is to add support for the new usernameAlias property
which provides an end-user readable description of the username.

Did some manual testing and unit tests pass